managed to create a mobile footer
