+++ lib/constant.dart
@@ -1,0 +1,8 @@
+import 'package:flutter/rendering.dart';
+
+const kFooterBackgroundColor = Color(0xFF27282c);
+
+const kHoverTextStyle = TextStyle(
+  color: Color(0xFF44C9F2),
+  letterSpacing: 1.25,
+);
