+++ lib/ui/screens/home_screen/mobile_home.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_explained_website/ui/screens/home_screen/hero_image.dart';
 import 'package:flutter_explained_website/ui/screens/home_screen/jumbotron.dart';
+import 'package:flutter_explained_website/widgets/footer/mobile_footer.dart';
 
 class MobileHome extends StatelessWidget {
   @override
@@ -9,6 +10,7 @@ class MobileHome extends StatelessWidget {
       children: [
         HeroImage(),
         Jumbotron(),
+        MobileFooter(),
       ],
     );
   }
